refactor: Generation module to start to seperate the MM logic from the rest and add some cleaner interfaces
